requirements.txt
aiogram==2.20
mysql-connector-python==8.0.30
